+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 ## Master
 - Allow nil GitHub milestone descriptions [@smalljd][] - [#253](https://github.com/danger/swift/pull/253)
 
+- Handle not existing PR author and contributor [@f-meloni][] - [#252](https://github.com/danger/swift/pull/252)
+
 ## 1.6.2
 
 - Update Architecture documentation [@mollyIV][] - [#250](https://github.com/danger/swift/pull/250)